+++ grpc/index.js
@@ -1,6 +1,7 @@
 const {defaultSocket} = require('./grpc_services');
 const emitSubscriptionError = require('./emit_subscription_error');
 const {grpcSslCipherSuites} = require('./grpc_services');
+const handleRemoveListener = require('./handle_remove_listener');
 const {maxReceiveMessageLength} = require('./grpc_services');
 const {packageTypes} = require('./grpc_services');
 const {protoFiles} = require('./grpc_services');
@@ -13,6 +14,7 @@ module.exports = {
   defaultSocket,
   emitSubscriptionError,
   grpcSslCipherSuites,
+  handleRemoveListener,
   maxReceiveMessageLength,
   packageTypes,
   protoFiles,