+++ lnd_methods/macaroon/index.js
@@ -1,6 +1,13 @@
 const getAccessIds = require('./get_access_ids');
 const grantAccess = require('./grant_access');
 const revokeAccess = require('./revoke_access');
+const subscribeToRpcRequests = require('./subscribe_to_rpc_requests');
 const verifyAccess = require('./verify_access');
 
-module.exports = {getAccessIds, grantAccess, revokeAccess, verifyAccess};
+module.exports = {
+  getAccessIds,
+  grantAccess,
+  revokeAccess,
+  subscribeToRpcRequests,
+  verifyAccess,
+};